Move ChangingShape.Multiblock to standalone test and add various compressors to it if they are available (blosc, mgard, zfp). Tests: Engine.BP.*/BPChangingShapeWithinStep.MultiBlock/*.
