+++ testing/adios2/engine/bp/CMakeLists.txt
@@ -130,6 +130,7 @@ bp_gtest_add_tests_helper(WriteReadLocalVariables MPI_ALLOW)
 bp_gtest_add_tests_helper(WriteReadLocalVariablesSel MPI_ALLOW)
 bp_gtest_add_tests_helper(WriteReadLocalVariablesSelHighLevel MPI_ALLOW)
 bp_gtest_add_tests_helper(ChangingShape MPI_ALLOW)
+bp_gtest_add_tests_helper(ChangingShapeWithinStep MPI_ALLOW)
 bp_gtest_add_tests_helper(WriteReadBlockInfo MPI_ALLOW)
 bp_gtest_add_tests_helper(WriteReadVariableSpan MPI_ALLOW)
 bp3_bp4_gtest_add_tests_helper(TimeAggregation MPI_ALLOW)
